// frontend/src/api/generated/models/TagSchema.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Galerie API
 * 本APIは、Galerieアプリケーションが画像やPDF化した漫画等のマルチメディアを提供するためのAPIである。  なお、このAPI自体はコンテンツそのものの配信を行わない。クライアントは本APIからコンテンツのメタデータを取得し、その中に含まれる情報を使ってコンテンツを取得する。
 *
 * The version of the OpenAPI document: 0.7.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TagDefinition } from './TagDefinition';
import {
    TagDefinitionFromJSON,
    TagDefinitionFromJSONTyped,
    TagDefinitionToJSON,
    TagDefinitionToJSONTyped,
} from './TagDefinition';
import type { TagSchemaVersion } from './TagSchemaVersion';
import {
    TagSchemaVersionFromJSON,
    TagSchemaVersionFromJSONTyped,
    TagSchemaVersionToJSON,
    TagSchemaVersionToJSONTyped,
} from './TagSchemaVersion';

/**
 * コンテンツの登録・検索・診断に適用する現在有効なタグスキーマ
 * @export
 * @interface TagSchema
 */
export interface TagSchema {
    /**
     *
     * @type {TagSchemaVersion}
     * @memberof TagSchema
     */
    version: TagSchemaVersion;
    /**
     * 定義されていないキーのタグを許可するか
     * @type {boolean}
     * @memberof TagSchema
     */
    allowAdditionalTags: boolean;
    /**
     * 必須タグの定義。定義がない場合は空配列
     * @type {Array<TagDefinition>}
     * @memberof TagSchema
     */
    required: Array<TagDefinition>;
    /**
     * 任意タグの定義。定義がない場合は空配列
     * @type {Array<TagDefinition>}
     * @memberof TagSchema
     */
    optional: Array<TagDefinition>;
}



/**
 * Check if a given object implements the TagSchema interface.
 */
export function instanceOfTagSchema(value: object): value is TagSchema {
    if (!('version' in value) || value['version'] === undefined) return false;
    if (!('allowAdditionalTags' in value) || value['allowAdditionalTags'] === undefined) return false;
    if (!('required' in value) || value['required'] === undefined) return false;
    if (!('optional' in value) || value['optional'] === undefined) return false;
    return true;
}

export function TagSchemaFromJSON(json: any): TagSchema {
    return TagSchemaFromJSONTyped(json, false);
}

export function TagSchemaFromJSONTyped(json: any, ignoreDiscriminator: boolean): TagSchema {
    if (json == null) {
        return json;
    }
    return {

        'version': TagSchemaVersionFromJSON(json['version']),
        'allowAdditionalTags': json['allowAdditionalTags'],
        'required': ((json['required'] as Array<any>).map(TagDefinitionFromJSON)),
        'optional': ((json['optional'] as Array<any>).map(TagDefinitionFromJSON)),
    };
}

export function TagSchemaToJSON(json: any): TagSchema {
    return TagSchemaToJSONTyped(json, false);
}

export function TagSchemaToJSONTyped(value?: TagSchema | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'version': TagSchemaVersionToJSON(value['version']),
        'allowAdditionalTags': value['allowAdditionalTags'],
        'required': ((value['required'] as Array<any>).map(TagDefinitionToJSON)),
        'optional': ((value['optional'] as Array<any>).map(TagDefinitionToJSON)),
    };
}